fixed errors took out Run, CostFunc and Bayesopt out of package
