+++ survivalnet/__init__.py
@@ -11,18 +11,17 @@ from . import model
 from .train import train
 
 # must be imported before Bayesian_Optimizaiton
-from .CostFunction import cost_func, aggr_st_cost_func, st_cost_func
+#from .CostFunction import cost_func, aggr_st_cost_func, st_cost_func
 
-from .Bayesian_Optimization import tune
+#from .Bayesian_Optimization import tune
 
-from .Run import Run
+#from .Run import Run
 
 
 # list out things that are available for public use
 __all__ = (
 
     # functions and classes of this package
-    'Run',
     'train',
 
     # sub-packages
