+++ survivalnet/optimization/GDLS.py
@@ -29,7 +29,7 @@ def _line_search_wolfe12(f, fprime, xk, pk, gfk, old_fval, old_old_fval,
                              **kwargs)
 
     if ret[0] is None:
-        print 'line search failed: try different one.'
+        #print 'line search failed: try different one.'
         with warnings.catch_warnings():
             warnings.simplefilter('ignore', LineSearchWarning)
             ret = line_search_wolfe2(f, fprime, xk, pk, gfk,
@@ -54,7 +54,7 @@ class GDLS(object):
         
         self.old_old_fval = None       
         N = self.theta_shape
-        print self.theta_shape
+        #print self.theta_shape
         
         self.theta = theano.shared(value=numpy.zeros(self.theta_shape, dtype=theano.config.floatX))
         self.theta.set_value(numpy.concatenate([e.get_value().ravel() for e in
@@ -118,10 +118,10 @@ class GDLS(object):
                  _line_search_wolfe12(f, fprime, self.theta_t, self.rho_t, self.gf_t,
                                       self.old_fval, self.old_old_fval, amin=1e-100, amax=1e100)
         except _LineSearchError:
-            print 'Line search failed to find a better solution.\n'         
+            #print 'Line search failed to find a better solution.\n'         
             theta_next = self.theta_t + self.gf_t * .0001
             return theta_next
-        print "Line Search Success! eps = ", self.eps_t
+        #print "Line Search Success! eps = ", self.eps_t
         theta_next = self.theta_t + self.eps_t * self.rho_t
         return theta_next
