+++ survivalnet/train.py
@@ -12,7 +12,7 @@ from optimization import isOverfitting
 
 def train(pretrain_set, train_set, test_set,
              pretrain_config, finetune_config, n_layers=10, n_hidden=140, coxphfit=False,
-             dropout_rate=0.5, non_lin=None, optim = 'GD', disp = True, earlystp = True, reg1 = 0, reg2 =0):    
+             dropout_rate=0.5, non_lin=None, optim = 'GD', disp = True, earlystp = True):    
     finetune_lr = theano.shared(numpy.asarray(finetune_config['ft_lr'], dtype=theano.config.floatX))
     learning_rate_decay = .989    
         
@@ -30,9 +30,7 @@ def train(pretrain_set, train_set, test_set,
         hidden_layers_sizes = [n_hidden] * n_layers,
         n_outs = 1,
         dropout_rate=dropout_rate,
-        non_lin=non_lin,
-	reg1 = reg1, 
-	reg2 = reg2)
+        non_lin=non_lin)
         
     #########################
     # PRETRAINING THE MODEL #
